+++ src/core/gameHelper.ts
@@ -1,8 +1,0 @@
-/// <reference path="./../reference.d.ts" />
-
-namespace Lightning {
-
-    export class EngineHelper {
-        
-    }
-}