+++ src/core/signals/signalManager.ts
@@ -9,126 +9,117 @@ namespace Lightning.Signals {
 
         private game:Engine;
         private _signals:{[name:string]:Signal};
-        /**
-         * signal manager constructor
-         * @param game
-         */
 
+        /**
+         * @description Signal Manager constructor
+         * 
+         * @param {engine} game
+         */
         constructor(game:Engine) {
             this.game = game;
             this._signals = {};
         }
 
-        getInsatance() {
-            
+        /**
+         * @description Create a new signal
+         * 
+         * @param {string} key
+         * 
+         * @returns {Signal}
+         */
+        create(key:string):Signal {
+            let signal = this._signals[key] = new Signal();
+            return signal || null;
         }
 
         /**
-         * create a new signal
-         * @param str
-         * @returns {any}
+         * @description Add a function to a signal
+         * 
+         * @param {string} key
+         * @param {function} fn
+         * @param {Object} listenerContext 
+         * 
+         * @returns {boolean}
          */
-        create(str:string) {
+        add(key:string, fn:Function, listenerContext:Object = null):boolean {
+            this.signal(key).add(fn, listenerContext);
+            return true;
+        }
+
+        /**
+         * @description Add a function to a signal only once
+         * 
+         * @param {string} key
+         * @param {function} fn
+         * @param {Object} listenerContext
+         * 
+         * @returns {boolean}
+         */
+        addOnce(key:string, fn:Function, listenerContext:Object):boolean {
+            this.signal(key).addOnce(fn, listenerContext);
+            return true;
+        }
+
+        /**
+         * @description Destroy the signal
+         * @param {string} key
+         * 
+         * @returns {boolean§}
+         */
+        destroy(key:string):boolean {
+            let s = this.signal(key);
+            s = null;
+            return true;
+        }
+
+        /**
+         * @description Change the active property on a signal
+         * 
+         * @param {string} key
+         * @param {boolean} active
+         * 
+         * @returns {boolean}
+         */
+        active(key:string, active:boolean):boolean {
             try {
-                this._signals[str] = new Signal();
-                return this._signals[str];
-            }
-            catch(e) {
-                console.error(e.message);
-                return null;
+                this.signal(key).active = active;
+            } catch(e) {
+                return false;
             }
         }
 
         /**
-         * add a function to the signal to fire on dispatch
-         * @param str
-         * @param fnct
-         * @param listenerContext? = null
+         * @description dispatch a signal and pass parameters
+         * 
+         * @param {string} key
+         * @param {Array} params
          */
-        add(str:string, fnct:Function, listenerContext:any = null) {
+        dispatch(key:string, ... params):boolean {
             try {
-                let s = this.signal(str);
-                this.signal(str).add(fnct,listenerContext);
-            }
-            catch(e) {
-                console.error(e.message);
+                this.signal(key).dispatch(params);
+                return true;
+            } catch(e) {
+                return false;
             }
         }
 
         /**
-         * add a function to the signal to fire only once on dispatch, then automatically destroy the function
-         * @param str
-         * @param fnct
+         * @description Returns a signal if it exists, else it will return null
+         * 
+         * @param {srting} key
+         * 
+         * @returns {Signal}
          */
-        addOnce(str:string, fnct:Function) {
-            try {
-                this.signal(str).addOnce(fnct);
-            }
-            catch(e) {
-                console.error(e.message);
-            }
+        signal(key:string):Signal {
+            return this.signal(key) || null;
         }
 
         /**
-         * destroy the entire signal
-         * @param str
-         */
-        destroy(str:string) {
-            try {
-                let s = this.signal(str);
-                s = null;
-            }
-            catch(e) {
-                console.error(e.message);
-            }
-        }
-
-        /**
-         * set the state of the signal (active, inactive)
-         * @param str
-         * @param val
-         */
-        active(str:string, val:boolean) {
-            try {
-                this.signal(str).active = val;
-            }
-            catch(e) {
-                console.error(e.message);
-            }
-        }
-
-        /**
-         * dispatch a signal with all the parameters
-         * @param str
-         * @param params
-         */
-        dispatch(str:string, ... params) {
-            try {
-                this.signal(str).dispatch(params);
-            }
-            catch(e) {
-                console.error(e.message);
-            }
-        }
-
-        /**
-         * return a signal
-         * @param str
-         * @returns {any}
-         */
-        signal(str:string) {
-            for(let i in this._signals) {
-                if(i === str) {
-                    return this._signals[i];
-                }
-            }
-            console.error('No signal exists with the key "' + str + '"');
-        }
-
-        /**
-         * check if signal is already created
-         * @param name
-         * @return boolean
+         * @description Return true if the signal is created, else return false
+         * 
+         * @param {string} name
+         * 
+         * @return {boolean}
          */
         has(name:string):boolean{
             return this._signals[name] !== undefined;