+++ src/core/states/state.ts
@@ -6,7 +6,12 @@ namespace Lightning {
         protected game:Engine;
         public loader:PIXI.loaders.Loader;
 
-        constructor(game:Engine, ...params) {
+        /**
+         * @description State constructor
+         * 
+         * @param {Engine} game
+         */
+        constructor(game:Engine) {
             super();
             this.game = game;
             this.loader = new PIXI.loaders.Loader();
@@ -15,41 +20,72 @@ namespace Lightning {
             this.loader.onComplete.add(this.preloadComplete, this);
         }
 
-        init(params) {
+        /**
+         * @description Initalization function
+         * 
+         * @param {Array} params
+         * 
+         * @returns {void}
+         */
+        init(params):void {
             this.preload();
         }
 
-        preload() {
+        /**
+         * @description Preload function. Used as a helper function to preload assets into the texture cache. Will skip and call the create function if there are no resources to load
+         * 
+         * @returns {void}
+         */
+        preload():void {
             if(Object.keys(this.loader.resources).length < 1) {
                 this.create();
             }
         }
 
-        create() {
+        /**
+         * @description Create function. Called after the preload function is complete or there is nothing to preload
+         * 
+         * @returns {void}
+         */
+        create():void {
 
-        }
-
-        update() {
-            
-        }
-
-        add(...params:Array<DisplayObject>) {
-            for(let i of params) {
-                this.addChild(i);
-            }
         }
 
         /**
-         * Called if the loader produces an error
+         * @description Update function. This is called by the state manager on every tick
          */
-        preloadError(err) {
+        update():void {
+            
+        }
+
+        /**
+         * @description Add children to this state. Helper functions should be migrated at some point
+         * 
+         * @returns {boolean}
+         */
+        add(...params:Array<DisplayObject>):boolean {
+            for(let i of params) {
+                this.addChild(i);
+            }
+
+            return true;
+        }
+
+        /**
+         * @description Called if the loader produces an error
+         * 
+         * @returns {void}
+         */
+        preloadError(err):void {
             console.log(err);
         }
 
         /**
-         * Called when a single file has completed loading
+         * @description Called when a single file has completed loading
+         * 
+         * @returns {void}
          */
-        preloadSingle(loader:PIXI.loaders.Loader, resource) {
+        preloadSingle(loader:PIXI.loaders.Loader, resource):void {
             // get the name of the loaded asset
             let file:string = resource.name;
             // remove the directory if you wish
@@ -59,9 +95,11 @@ namespace Lightning {
         }
 
         /**
-         * Called when the loader has finished loading everything
+         * @description Called when the loader has finished loading everything
+         * 
+         * @returns {void}
          */
-        preloadComplete(resources) {
+        preloadComplete(resources):void {
             this.create();
         }
     }