+++ src/core/tween-generator/Interfaces/Property.ts
@@ -1,4 +1,0 @@
-declare interface Property {
-    prop:string;
-    val:number;
-}