+++ src/reference.d.ts
@@ -11,6 +11,8 @@
 /// <reference path ="./interfaces/iRange.ts" />
 /// <reference path ="./interfaces/iPointRange.ts" />
 /// <reference path ="./interfaces/iStateMap.ts" />
+/// <reference path ="./interfaces/iTweenCallback.ts" />
+/// <reference path ="./interfaces/iTweenProperty.ts" />
 /**
  * Utils
  */
@@ -19,7 +21,8 @@
 /**
  * Core
  */
-/// <reference path="./core/game.ts" />
+/// <reference path="./core/engine.ts" />
+/// <reference path="./core/engineHelper.ts" />
 /// <reference path="./core/states/state.ts" />
 /// <reference path="./core/states/stateManager.ts" />
 /// <reference path="./core/physics/physicsManager.ts" />
@@ -41,14 +44,14 @@
 /// <reference path="./display/geometry.ts" />
 /// <reference path="./display/hitarea.ts" />
 /// <reference path="./display/button.ts" />
-/// <reference path="./display/particle.ts" />
-/// <reference path="./display/particleEmitter.ts" />
-/// <reference path="./display/parallax.ts" />
-/// <reference path="./core/tween-generator/Easing.ts" />
-/// <reference path="./core/tween-generator/Events.ts" />
-/// <reference path="./core/tween-generator/Frame.ts" />
-/// <reference path="./core/tween-generator/Tween.ts" />
-/// <reference path="./core/tween-generator/TweenManager.ts" />
+/// <reference path="./particleEmitter/particle.ts" />
+/// <reference path="./particleEmitter/particleEmitter.ts" />
+/// <reference path="./parallax/parallax.ts" />
+/// <reference path="./tweens/Easing.ts" />
+/// <reference path="./tweens/Events.ts" />
+/// <reference path="./tweens/Frame.ts" />
+/// <reference path="./tweens/Tween.ts" />
+/// <reference path="./tweens/TweenManager.ts" />
 /// <reference path="./core/signals/signal.ts" />
 /// <reference path="./core/signals/signalBinding.ts" />
 /// <reference path="./core/signals/signalManager.ts" />